requirements.txt
spotipy==2.17.1
flask
openai
python-dotenv
flask_session
